refactor(macos): remove macOS Mojave patch support
Remove the macOS Mojave patch support that was added to fix the initial teething issues with Apple's
new OS and Xcode releases. This functionality was already disabled by default and should not be
needed on up-to-date Mojave installations using the latest Xcode versions.

BREAKING CHANGE: macOS Mojave with outdated signed SDK header packages or outdated Xcode versions
are not supported.