feat(linux): install with Homebrew on Linux
The main purpose of this feature is to allow testing of the Homebrew
installation on my Windows and Linux development environments inside
Ubuntu and Debian Docker containers.

This should be considered experimental and not the preferred way to
install pyenv on Linux as it comes with a few quirks.

To address known issue with pyenv on Homebrew on Linux, the Homebrew
readline package is automatically uninstalled and then reinstalled
while building Python versions with this role.